third_party/WebKit/LayoutTests/imported/web-platform-tests/html/browsers/the-window-object/window-prototype-chain-expected.txt

This is a testharness.js-based test.
FAIL window object assert_equals: expected "[object Window]" but got "[object global]"
FAIL Window.prototype assert_equals: expected "[object WindowPrototype]" but got "[object Object]"
FAIL Global scope polluter assert_equals: expected "[object WindowProperties]" but got "[object Object]"
FAIL EventTarget.prototype assert_equals: expected "[object EventTargetPrototype]" but got "[object Object]"
PASS Object.prototype 
Harness: the test ran to completion.
